pyproject.toml
[tool.poetry]
name = "audio-ai"
version = "0.1.0"
description = ""
authors = ["Kevin Weingarten <weingarten.kevin@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
torchaudio = "0.12.1"
torchvision = "0.13.1"
pandas = "^1.5.2"
torchsummary = "^1.5.1"
python-dotenv = "^0.21.0"
tensorflow = "^2.11.0"
keras = "^2.11.0"
beautifulsoup4 = "^4.11.1"
lxml = "^4.9.2"
pyspark = "^3.3.1"



[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
